Refactor database imports and add fields API routes
