+++ server.py
@@ -1,9 +1,10 @@
-
 from fastapi import FastAPI
 from routes.teams import router as teams_router
+from routes.fields import router as fields_router
 
 app = FastAPI()
 app.include_router(teams_router)
+app.include_router(fields_router)
 
 if __name__ == "__main__":
     import uvicorn
